requirements.txt
altair_saver==0.5.0
altair_viewer==0.4.0
AppKit==0.2.8
atheris==2.2.2
beautifulsoup4==4.11.1
bjoern==3.2.2
boto3==1.26.124
botocore==1.29.124
brotlicffi==1.0.9.2
brotlipy==0.7.0
cftime==1.6.2
chart_studio==1.1.0
Cheetah==2.4.4
cherrypy==18.8.0
cloudpickle==2.0.0
ConfigParser==5.3.0
cryptography==39.0.1
ctypes_snappy==1.03
cycler==0.11.0
Cython==0.29.34
cytoolz==0.12.0
dask==2022.7.0
defusedxml==0.7.1
diesel==3.0.24
Django==4.2
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
eventlet==0.33.3
fastparquet==2023.4.0
flup==1.0.3
Foundation==0.1.0a0.dev1
fqdn==1.5.1
fsspec==2022.11.0
genapi==0.0.8
gevent==22.10.2
gitdb_speedups==0.1.0
gunicorn==20.1.0
HTMLParser==0.0.2
hypothesis==6.75.1
ipython==8.10.0
ipywidgets==7.6.5
isoduration==20.11.0
jnius==1.1.0
JPype1==1.4.1
jsonpointer==2.3
keyring==23.4.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.4
matplotlib==3.7.1
meinheld==1.0.2
mtrand==0.1
mypy==1.2.0
nltk==3.7
nose==1.3.7
numarray==1.5.1
Numeric==24.2
numexpr==2.8.4
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.10
PasteScript==3.3.0
pickle5==0.0.12
pkgutil_resolve_name==1.3.10
plotly==5.9.0
psutil==5.9.0
psycopg2==2.9.6
py==1.11.0
pycares==4.3.0
pycurl==7.45.2
PyInstaller==5.10.1
pymysql==1.0.3
pyobjc_framework_Cocoa==9.0
pyOpenSSL==23.0.0
pyOpenSSL==23.1.1
pyperf==2.6.0
PyQt4==4.11.4
PyQt5==5.15.9
PyQt6==6.5.0
pyrsistent==0.18.0
PySide2==5.15.2.1
PySide6==6.5.0
pytest==7.1.2
python-socketio==5.8.0
pyxlsb==1.0.10
PyYAML==6.0
PyYAML==6.0
QtPy==2.2.0
railroad==0.5.0
recommonmark==0.7.1
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rocket==1.2.8
s3fs==2023.4.0
scikit_learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
sets==0.3.2
setuptools_scm==7.1.0
simplejson==3.19.1
Sphinx==5.0.2
SQLAlchemy==1.4.39
statsmodels==0.13.5
sympy==1.11.1
tables==3.7.0
testbench==0.1.2
threadpoolctl==3.1.0
torch==1.12.1
traitlets==5.9.0
trove_classifiers==2023.4.29
truststore==0.7.0
uri_template==1.2.0
urllib3_secure_extra==0.1.0
vega_datasets==0.9.0
waitress==2.1.2
watchdog==2.1.6
webcolors==1.13
wordcloud==1.8.2.2
wordcloud==1.8.1
xarray==2022.11.0
xlrd==2.0.1
xlsxwriter==3.1.0
xlwt==1.3.0
xmlrpclib==1.0.1
